add UI enhancement and animations

+++ TodoList App/View/ListView.swift
@@ -12,20 +12,30 @@ struct ListView: View {
     @EnvironmentObject var listViewModel:ListViewModel
     
     var body: some View {
-        List{
-            ForEach(listViewModel.items){item in
-                
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear){
-                            listViewModel.updateItem(item:item)
-                        }
-                    }
+        ZStack{
+            if listViewModel.items.isEmpty{
+                NoItemView()
+                    .transition(AnyTransition.opacity
+                        .animation(.easeIn))
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
+            else{
+                List{
+                    ForEach(listViewModel.items){item in
+                        
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear){
+                                    listViewModel.updateItem(item:item)
+                                }
+                            }
+                    }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
+                .listStyle(PlainListStyle())
+            }
         }
-        .listStyle(PlainListStyle())
+        
         .navigationTitle("Todo List📝")
         .navigationBarItems(
             leading: EditButton(),
